Add Enzyme, move Redux files into ./src/store directory, fix import paths
