+++ react-ui/package.json
@@ -30,7 +30,8 @@
     "eject": "react-scripts eject"
   },
   "devDependencies": {
-    "eslint-config-airbnb": "^17.1.0",
-    "react-test-renderer": "^16.6.3"
+    "enzyme": "^3.7.0",
+    "enzyme-adapter-react-16": "^1.7.0",
+    "eslint-config-airbnb": "^17.1.0"
   }
 }
