+++ react-ui/src/store/reducers/rootReducer.js
@@ -1,7 +1,7 @@
 import { combineReducers } from 'redux';
 
 // Reducers
-import testReducer from './TestReducer';
+import testReducer from './testReducer';
 
 export default combineReducers({
   testReducer,
