+++ react-ui/src/tests/App.test.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import { Provider } from 'react-redux';
-import configureStore from '../store';
+import configureStore from '../store/store';
 import App from '../containers/App/App';
 
 it('renders without crashing', () => {
